Refactor tar_gz tests into separate file
